Fixed CI error, take 2

+++ source/core/mortar_coupling_manager.cc
@@ -792,9 +792,9 @@ compute_minimum_height(const Triangulation<dim>      &triangulation,
       // Direction of the rotation axis
       unsigned int   direction = 0;
       Tensor<1, dim> axis      = mortar_parameters.rotation_axis;
-      for (unsigned int d = 1; d < dim; ++d)
+      for (unsigned int d = 1; d < static_cast<unsigned int>(dim); ++d)
         if (std::abs(axis[d]) > std::abs(axis[direction]))
-          direction = static_cast<unsigned int>(d);
+          direction = d;
 
       AssertThrow(
         std::abs(axis[direction]) > 0.99,
